refactor so cli code contained in one folder
