+++ memgpt/main.py
@@ -30,7 +30,9 @@ from memgpt.persistence_manager import (
     InMemoryStateManagerWithPreloadedArchivalMemory,
     InMemoryStateManagerWithFaiss,
 )
-
+from memgpt.cli.cli import run
+from memgpt.cli.cli_config import configure
+from memgpt.cli.cli_load import app as load_app
 from memgpt.config import Config, MemGPTConfig, AgentConfig
 from memgpt.constants import MEMGPT_DIR
 from memgpt.connectors import connector
@@ -43,9 +45,10 @@ from memgpt.openai_tools import (
 import asyncio
 
 app = typer.Typer()
-metadata_app = typer.Typer()
-app.add_typer(connector.app, name="load")
-app.add_typer(metadata_app, name="list")
+app.command(name="run")(run)
+app.command(name="configure")(configure)
+# load data commands
+app.add_typer(load_app, name="load")
 
 
 def clear_line():
@@ -116,263 +119,6 @@ def load(memgpt_agent, filename):
         print(f"Loaded persistence manager from {filename}")
     except Exception as e:
         print(f"/load warning: loading persistence manager from {filename} failed with: {e}")
-
-
-@metadata_app.command("agents")
-def list_agents():
-    """List all agents"""
-    table = PrettyTable()
-    table.field_names = ["Name", "Model", "Persona", "Human", "Data Source"]
-    for agent_file in utils.list_agent_config_files():
-        agent_name = os.path.basename(agent_file).replace(".json", "")
-        agent_config = AgentConfig.load(agent_name)
-        table.add_row([agent_name, agent_config.model, agent_config.persona, agent_config.human, agent_config.data_source])
-    print(table)
-
-
-@metadata_app.command("humans")
-def list_humans():
-    """List all humans"""
-    table = PrettyTable()
-    table.field_names = ["Name", "Text"]
-    for human_file in utils.list_human_files():
-        name = os.path.basename(human_file)
-        text = humans.get_human_text(name)
-        table.add_row([name, text])
-    print(table)
-
-
-@metadata_app.command("personas")
-def list_personas():
-    """List all personas"""
-    table = PrettyTable()
-    table.field_names = ["Name", "Text"]
-    for persona_file in utils.list_persona_files():
-        name = os.path.basename(persona_file)
-        text = personas.get_persona_text(name)
-        table.add_row([name, text])
-    print(table)
-
-
-@metadata_app.command("sources")
-def data_sources(data):
-    """List all data sources"""
-    table = PrettyTable()
-    table.field_names = ["Name", "Create Time", "Agents"]
-    for data_source_file in os.listdir(os.path.join(MEMGPT_DIR, "archival")):
-        name = os.path.basename(data_source_file)
-        table.add_row([name, "TODO", "TODO"])
-    print(table)
-
-
-@app.command()
-def add(
-    option: str,  # [human, persona]
-    name: str = typer.Option(help="Name of human/persona"),
-    text: str = typer.Option(None, help="Text of human/persona"),
-    filename: str = typer.Option(None, "-f", help="Specify filename"),
-):
-    """Add a person/human"""
-
-    if option == "persona":
-        directory = os.path.join(MEMGPT_DIR, "personas")
-    elif option == "human":
-        directory = os.path.join(MEMGPT_DIR, "humans")
-    else:
-        raise ValueError(f"Unknown kind {kind}")
-
-    if filename:
-        assert text is None, f"Cannot provide both filename and text"
-        # copy file to directory
-        shutil.copyfile(filename, os.path.join(directory, name))
-    if text:
-        assert filename is None, f"Cannot provide both filename and text"
-        # write text to file
-        with open(os.path.join(directory, name), "w") as f:
-            f.write(text)
-
-
-@app.command()
-def configure():
-    """Updates default MemGPT configurations"""
-
-    default_provider = "openai"
-
-    # openai credentials
-    use_openai = questionary.confirm("Do you want to enable MemGPT with Open AI?").ask()
-    if use_openai:
-        # search for key in enviornment
-        openai_key = os.getenv("OPENAI_API_KEY")
-        if not openai_key:
-            openai_key = questionary.text("Open AI keys not found in enviornment - please enter:").ask()
-        default_openai = questionary.confirm("Use OpenAI as default provider?").ask()
-        if default_openai:
-            default_provider = "openai"
-
-    # azure credentials
-    use_azure = questionary.confirm("Do you want to enable MemGPT with Azure?").ask()
-    use_azure_deployment_ids = False
-    if use_azure:
-        # search for key in enviornment
-        azure_key = os.getenv("AZURE_API_KEY")
-        azure_endpoint = (os.getenv("AZURE_ENDPOINT"),)
-        azure_version = (os.getenv("AZURE_VERSION"),)
-        azure_deployment = (os.getenv("AZURE_OPENAI_DEPLOYMENT"),)
-        azure_embedding_deployment = os.getenv("AZURE_OPENAI_EMBEDDING_DEPLOYMENT")
-
-        if all([azure_key, azure_endpoint, azure_version]):
-            print(f"Using Microsoft endpoint {azure_endpoint}.")
-            if all([azure_deployment, azure_embedding_deployment]):
-                print(f"Using deployment id {azure_deployment}")
-                use_azure_deployment_ids = True
-            default_azure = questionary.confirm("Use Azure as default provider?").ask()
-            if default_azure:
-                default_provider = "azure"
-        else:
-            print("Missing enviornment variables for Azure. Please set then run `memgpt configure` again.")
-            # TODO: allow for manual setting
-            use_azure = False
-
-    # TODO: configure local model
-
-    # default model
-    model_options = []
-    if use_openai:
-        model_options += ["gpt-3.5-turbo", "gpt-3.5", "gpt-4"]
-    default_model = questionary.select(
-        "Select default model (recommended: gpt-4):", choices=["gpt-3.5-turbo", "gpt-3.5", "gpt-4"], default="gpt-4"
-    ).ask()
-
-    # defaults
-    personas = [os.path.basename(f).replace(".txt", "") for f in utils.list_persona_files()]
-    print(personas)
-    default_persona = questionary.select("Select default persona:", personas, default="sam_pov").ask()
-    humans = [os.path.basename(f).replace(".txt", "") for f in utils.list_human_files()]
-    print(humans)
-    default_human = questionary.select("Select default human:", humans, default="cs_phd").ask()
-
-    # TODO: figure out if we should set a default agent or not
-    default_agent = None
-    # agents = [os.path.basename(f).replace(".json", "") for f in utils.list_agent_config_files()]
-    # if len(agents) > 0: # agents have been created
-    #    default_agent = questionary.select(
-    #        "Select default agent:",
-    #        agents
-    #    ).ask()
-    # else:
-    #    default_agent = None
-
-    # TODO: allow configuring embedding model
-
-    config = MemGPTConfig(
-        model=default_model,
-        provider=default_provider,
-        default_persona=default_persona,
-        default_human=default_human,
-        default_agent=default_agent,
-        openai_key=openai_key if use_openai else None,
-        azure_key=azure_key if use_azure else None,
-        azure_endpoint=azure_endpoint if use_azure else None,
-        azure_version=azure_version if use_azure else None,
-        azure_deployment=azure_deployment if use_azure_deployment_ids else None,
-        azure_embedding_deployment=azure_embedding_deployment if use_azure_deployment_ids else None,
-    )
-    print(f"Saving config to {config.config_path}")
-    config.save()
-
-
-@app.command()
-def run(
-    persona: str = typer.Option(None, help="Specify persona"),
-    agent: str = typer.Option(None, help="Specify agent save file"),
-    human: str = typer.Option(None, help="Specify human"),
-    model: str = typer.Option(None, help="Specify the LLM model"),
-    data_source: str = typer.Option(None, help="Specify data source to attach to agent"),
-    first: bool = typer.Option(False, "--first", help="Use --first to send the first message in the sequence"),
-    debug: bool = typer.Option(False, "--debug", help="Use --debug to enable debugging output"),
-    no_verify: bool = typer.Option(False, "--no_verify", help="Bypass message verification"),
-):
-
-    """Start chatting with an MemGPT agent
-
-    Example usage: `memgpt run --agent myagent --data-source mydata --persona mypersona --human myhuman --model gpt-3.5-turbo`
-
-    :param persona: Specify persona
-    :param agent: Specify agent name (will load existing state if the agent exists, or create a new one with that name)
-    :param human: Specify human
-    :param model: Specify the LLM model
-    :param data_source: Specify data source to attach to agent (if new agent is being created)
-
-    """
-
-    # setup logger
-    utils.DEBUG = debug
-    logging.getLogger().setLevel(logging.CRITICAL)
-    if debug:
-        logging.getLogger().setLevel(logging.DEBUG)
-
-    # if no config, run configure
-    if not MemGPTConfig.exists():
-        print("No config found, running configure...")
-        configure()
-
-    # load config defaults
-    config = MemGPTConfig.load()
-
-    # override with command line arguments
-    if debug:
-        config.debug = debug
-    if no_verify:
-        config.no_verify = no_verify
-
-    # create agent config
-    if agent and AgentConfig.exists(agent):  # use existing agent
-        print(f"Using existing agent {agent}")
-        agent_config = AgentConfig.load(agent)
-        print("State path:", agent_config.save_state_dir())
-        print("Persistent manager path:", agent_config.save_persistence_manager_dir())
-        print("Index path:", agent_config.save_agent_index_dir())
-        # persistence_manager = LocalStateManager(agent_config).load() # TODO: implement load
-        # TODO: load prior agent state
-        assert not any(
-            [persona, human, model]
-        ), f"Cannot override existing agent state with command line arguments: {persona}, {human}, {model}"
-
-        # load existing agent
-        memgpt_agent = AgentAsync.load_agent(memgpt.interface, agent_config)
-    else:  # create new agent
-        # create new agent config: override defaults with args if provided
-        agent_config = AgentConfig(
-            name=agent if agent else None,
-            persona=persona if persona else config.default_persona,
-            human=human if human else config.default_human,
-            model=model if model else config.model,
-        )
-
-        # attach data source to agent
-        agent_config.attach_data_source(data_source)
-
-        # TODO: allow configrable state manager (only local is supported right now)
-        persistence_manager = LocalStateManager(agent_config)  # TODO: insert dataset/pre-fill
-
-        # save new agent config
-        agent_config.save()
-        print(f"Created new agent {agent_config.name}")
-
-        # create agent
-        memgpt_agent = presets.use_preset(
-            presets.DEFAULT,
-            agent_config,
-            agent_config.model,
-            agent_config.persona,
-            agent_config.human,
-            memgpt.interface,
-            persistence_manager,
-        )
-
-    # start event loop
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(run_agent_loop(memgpt_agent, first, False, config))  # TODO: add back no_verify
 
 
 # @app.callback(invoke_without_command=True)  # make default command
